src/requirements.txt
Flask
Flask-RESTx
SQLAlchemy
mysql.connector
pymysql
numpy
pandas
gunicorn
psutil
cachetools